Clean up VEXM pattern-matching in the x86 decoder
This enables general pattern-matching over the VEXM type in the
"evaluate" function. (It's a bit ad hoc to do it this way for the
different types rather than via general code, but we will probably
make more substantial changes here soon.) Taking advantage of
this improvement, we rewrite decode_aux with pattern-matching
instead of the nested conditional branches for the VEXM part,
which seems more natural.