Regenerate ML-KEM forward and inverse NTT with SLOTHY
This reruns the generation of the optimized code from the clean
forms in mlkem-native using a slight variant of the script here:

  https://github.com/pq-code-package/mlkem-native/blob/main/mlkem/native/aarch64/src/optimize.sh

The microarchitectural target is Neoverse N1, versus the Cortex A-55
currently used in mlkem-native ("Arm_Neoverse_N1_experimental"), e.g.

  slothy-cli Arm_AArch64 $TARGET \
    ntt_clean.S -o ntt_opt.S \
    -r ntt_asm_clean,ntt_asm_opt \
    -l layer012_start \
    -l layer3456_start \
    -c sw_pipelining.enabled=true \
    -c inputs_are_outputs \
    -c reserved_regs="[x18--x30,sp]" \
    -c sw_pipelining.minimize_overlapping=False \
    -c variable_size \
    -c constraints.stalls_first_attempt=64

This is followed by the same miscellaneous tweaks to make the code
better fit the current s2n-bignum proof setup (e.g. writing away
LD1 and ST4).