Update ML-KEM forward NTT code in line with mlkem-native ST4 change
This adopts as a starting-point the newly modified code from the
mlkem-native repository (commit 026de4a)

  https://github.com/pq-code-package/mlkem-native/blob/main/mlkem/native/aarch64/src/ntt_clean.S

It's then re-slothied for Neoverse N1 by:

  slothy-cli Arm_AArch64 Arm_Neoverse_N1_experimental \
    ntt_clean.S -o ntt_opt.S \
    -r ntt_asm_clean,ntt_asm_opt \
    -l layer012_start \
    -l layer3456_start \
    -c sw_pipelining.enabled=true \
    -c inputs_are_outputs \
    -c reserved_regs="[x18--x30,sp]" \
    -c sw_pipelining.minimize_overlapping=False \
    -c variable_size \
    -c constraints.stalls_first_attempt=64

The result is tweaked to match up with s2n-bignum again, but there is
now no need to eliminate ST4 post-hoc because the change above already
does it for the pre-SLOTHY clean code. This is not only more elegant
but seems markedly more efficient on both Graviton2 (2.5% better) and
Graviton3 (6% better).