Switch to the `*_neon` functions
This patch makes the `*_neon` functions replace their original scalar
implementations. This (partially) resolves the divergence between the
set of functions supported in Arm and in x86. There are still a few
functions that are diverged - `bignum_emontredc_8n_cdiff` and
`bignum_copy_row_from_table_*` which only exists in Arm - but all other
functions are converged into one.

The original scalar functions are moved to the `unopt/` directories.
Their proofs are merged into the `*_neon.ml` proofs, which are again
renamed to the original `*.ml`.
All `_NEON` and `_neon` suffixes are removed.

Also, this patch applies the NIST P-256 optimized field operations to
`p256_scalarmulbase` which was missing in the past.