Extend ARM cosimulator to cover some memory operations
This enables cosimulation of LD1/ST1/LD2/ST2 in the simulation setup
as well as the existing register-to-register instructions. Adding more
memory operations should now be relatively routine, although some
individual coding is needed to set up self-contained stack-safe test
sequences, as done for the present operations in "cosimulate_ldst_12()".
The main changes to support memory operations in the simulator are:

- Factor out the main cosimulation code from the random selection of
  instructions, and also generalize cosimulation to accept any list
  of instructions, not just a single one.

- Save and check additional state during cosimulation, a 256-byte buffer
  at SP,...,SP+255. The stack pointer value must be the same before and
  after, and the results in other registers independent of that value.

- Complicate the tactics a bit in the memory case to re-express memory
  assignments in a way that makes sure they are applicable, in the
  final resort reducing everything to byte arrays.

- For memory operations, set up a self-contained test harness that
  satisfies the constraints above, i.e. maintains original stack pointer
  and does not write outside [SP],...,[SP+255], but still tests the
  memory operation over a good range of instances.

There are a few other minor changes that make the code slightly more
portable, and in particular make it run cleanly on Mac OS using Clang
there is now a "_harness" as well as "harness" entry point name for
the test harness, and the printf-ing of uint64_t uses more strictly
portable macros.

At the moment, about 10% of runs are the memory operations and 90% are
the original register-to-register instructions. The latter should work
in much the same way, with only a slight slowdown from the additional
infrastructure.