Add microarchitectural events for Arm
This patch adds microarchitectural events for Arm.
Even if this is not officially a component of an architectural state,
This is necessary to describe the safety of assembly programs from
side-channel attacks.
We define that an instruction raises a microarchitectural event if
its cycles/power consumption/anything that can be observed by a
traditional side-channel attacker can vary depending on the inputs of
the instruction.
Its kinds (EventLoad/Store/...) describe the events distinguishable from
each other by the attacker, and their parameters describe the values
that are inputs and/or outputs of the instructions that will affect the
observed cycles/etc.
One instruction can raise multiple events (e.g., one that reads PC from
the memory and jumps to the address, even though this case will not exist
in Arm)

This largely imports from the codebase of Abdal in his last summer internship,
but reinforces the events to include the actual opcode that has been executed
(since in general it can affect the microarchitecture statues, e.g., branch predictor).
The opcode will be 4-byte bitvector for Arm since one instruction has 4 bytes.
Also, the branch event has all source and target PCs.

This implies that the non-interference reasoning on the event list is
slightly more complicated (it will include extra `word_add PC ...`),
but dealing with such patterns will not be too complicated I am hoping.