Add microarchitectural events for Arm
This patch adds microarchitectural events for Arm.
Even if this is not officially a component of an architectural state,
This is necessary to describe the safety of assembly programs from
side-channel attacks.
We define that an instruction raises a microarchitectural event if
its cycles/power consumption/anything that can be observed by a
traditional side-channel attacker can vary depending on the inputs of
the instruction.
Its kinds (EventLoad/Store/...) describe the events distinguishable from
each other by the attacker, and their parameters describe the values
that are inputs and/or outputs of the instructions that will affect the
observed cycles/etc.
One instruction can raise multiple events (e.g., one that reads PC from
the memory and jumps to the address, even though this case will not exist
in Arm)

This largely imports from the codebase of Abdal in his last summer internship.
There is one important difference that I added: the event on conditional
jump isn't taking its boolean condition anymore, but now the destination PC
of the jumps.
This is because Arm has a jump instruction to the PC stored in a register
(which is `RET` but we also have `arm_BL_ABSOLUTE`).
This implies that the non-interference reasoning on the event list is
slightly more complicated (it will include extra `word_add PC ...`),
but its pattern will not be too complicated I am hoping.