Patch up some crude SHA3 restrictions for ARM build and test
This makes a few changes to accommodate the fact that the ARM
mlkem_keccak_f1600_alt code relies on sha3 instructions that are
not in all aarch64 versions.

The builds now always add the flags "-march=armv8.2-a+sha3" to the
assembler to make sure the code is at least assembled regardless of
platform.

On the other hand, the test and benchmarking code attempts to restrict
execution of mlkem_keccak_f1600_alt to platforms supporting the sha3
instructions. This is determined in the tests/arch.h file in a rather
crude way: for Linux it is done properly using a standard interface,
while for non-Linux platforms we assume (1) Apple platforms all have
sha3, and (2) other platforms don't. It would be much better to have
a clean portable solution but this seems to be non-trivial. Note that
the solution used on x86 to determine BMI (reading system registers
via inline asm) is not available on ARM since the analogous system
register read (bits 32-35 of ID_AA64ISAR0_EL1) does not execute in
unprivileged code. See

  https://stackoverflow.com/questions/60071590/reading-armv8-a-registers-with-devmem-from-gnu-linux-shell