Add 2-way batch Keccak code based on mlkem-native
This adds a 2-way batch implementation of Keccak-f1600, taking
a 50-element input/output array considered as two successive
independent Keccak states a[0..24] and a[25..49]; the same
Keccak-f1600 permutation is applied to each one. The code is
closely based on an mlkem-native original

  https://github.com/pq-code-package/mlkem-native/blob/main/mlkem/fips202/native/aarch64/src/keccak_f1600_x2_v84a_asm_clean.S

The main change is the use of transpositions and plain load/store
in place of ld2/st2.