Add 4-way batch hybrid Keccak code based on mlkem-native
This adds a 4-way batch implementation of Keccak-f1600, taking a
100-element input/output array considered as four successive
independent Keccak states a[0..24], a[25..49], a[50..74], a[75..99];
the same Keccak-f1600 permutation is applied to each one. The code is
closely based on an mlkem-native original that mixes SIMD and scalar
operations:

      https://github.com/pq-code-package/mlkem-native/blob/main/mlkem/fips202/native/aarch64/src/keccak_f1600_x4_scalar_v84a_asm_hybrid_opt.S

The main change is the use of transpositions and plain load/store
in place of ld2/st2.