Add ENDBR64 to x86 model
This is intended to support the insertion of ENDBR64 into the x86
code to support indirect branch tracking (IBT), part of CET
(control-flow enforcement technology).

The actual semantics assumed is simply a NOP; on machines without
CET enabled, including older machines from before the feature
existed, this is indeed how it behaves. We do not attempt to model
the restrictions that are imposed when CET is enabled, and such
restrictions arguably belong on the indirect branches themselves, of
which there are none in s2n-bignum itself.

For more discussion see

  https://stackoverflow.com/questions/56905811/what-does-the-endbr64-instruction-actually-do