Enable IBT by default in the code
This makes the code compile with IBT unless explicitly flagged
otherwise in the compiler options, hence inserting one initial
ENDBR64 instruction at the main entry point for each function.

The copy of the code given in each "define_assert_from_elf" is
also updated correspondingly.

So far, the proofs are not updated, though the