Add support for UMLAL2, UMLSL{2}, SMULL{2}, SMLAL{2}, SMLSL{2}
This commit extends the AArch64 architecture model with support
for `UMLAL2, UMLSL{2}, SMULL{2}, SMLAL{2}, SMLSL{2}`.

This follows the existing templates for UMULL and UMLAL. The difference
in encoding between Sxxx and Uxxx is bit-29 only, which is 1 for unsigned
and 0 for signed variants. Similarly, xMLAL and xMLSL only differ in bit 13.

Semantically, we merely need to switch from zero-extend to sign-extend,
and replace word_add by word_sub for xMLSL variants.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>