AArch64/SIMD: Add negative decoding tests for invalid sizes
Various NEON instructions have a 2-bit size parameter but are
only valid for the values {0b00, 0b01, 0b10}. Notably, this
is the case for all SIMD integer multiplication instructions.

For those instructions, `simulator_iclasses.ml` previously
listed only the valid encodings. This would test that the functional
description of valid encodings was correct, but it would not test
that the architecture and the model agreed on them being invalid.

This commit relaxes the encoding patterns in `simulator_iclasses.ml`
so that invalid size encodings are considered as well.

A similar change may be possible for immediates required to be
non-zero, but the architecture is not clear on what decoding does
with them: It says `SEE(asimdimm)` rather than `UNDEF`. Those
are therefore not touched for now.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>