Enable x86 indirect branch tracking in code and proofs (#203)
* x86: add script for adding endbr{32,64} instructions

* Implement CET for amd64

* Sync AT&T syntax files

* Add ENDBR64 to x86 model

This is intended to support the insertion of ENDBR64 into the x86
code to support indirect branch tracking (IBT), part of CET
(control-flow enforcement technology).

The actual semantics assumed is simply a NOP; on machines without
CET enabled, including older machines from before the feature
existed, this is indeed how it behaves. We do not attempt to model
the restrictions that are imposed when CET is enabled, and such
restrictions arguably belong on the indirect branches themselves, of
which there are none in s2n-bignum itself.

For more discussion see

  https://stackoverflow.com/questions/56905811/what-does-the-endbr64-instruction-actually-do

* Enable IBT by default in the code

This makes the code compile with IBT unless explicitly flagged
otherwise in the compiler options, hence inserting one initial
ENDBR64 instruction at the main entry point for each function.

The copy of the code given in each "define_assert_from_elf" is
also updated correspondingly.

So far, the proofs are not updated, though the

* Re-derive non-IBT version of code and restore proofs

This uses the function define_trimmed that trims away the first 4
bytes of the code (expected to be ENDBR64) and makes a definition
matching the non-IBT original, hence restoring the proofs.

* Use symbolic lengths for most x86 toplevel correctness theorems

This switches most xxx_SUBROUTINE_CORRECT statements to use
explicit "LENGTH code" hypotheses in place of explicit numbers.
The main unchanged ones are those with data in the text section
as well.

* Add IBT variants of x86 toplevel correctness statements

Every xxx_SUBROUTINE_CORRECT result, both standard and Windows ABI,
now has a variant xxx_IBT_SUBROUTINE_CORRECT in the file, with a
proof that mostly works automatically via the new ADD_IBT_RULE,
corresponding to the case where the code is prefixed by ENDBR64.
This rule fails to handle the cases with data in the text section;
this occurs in these proofs that will need fixing next:

        edwards25519_scalarmuldouble_alt.ml
        edwards25519_scalarmuldouble.ml
        edwards25519_scalarmulbase_alt.ml
        edwards25519_scalarmulbase.ml
        curve25519_x25519base_alt.ml
        curve25519_x25519base.ml

* Fix remaining IBT proof breaks

This is done by making ADD_IBT_RULE a little more robust so that it
handles the case of appended code and data tables, and also modifies
explicit code length numbers as well as symbolic lengths.

* Invert naming conventions for x86 Windows and IBT versions

This switches around the naming conventions in correctness theorems
and machine code definitions:

 * Instead of the IBT forms of correctness theorems having their names
   flagged with _IBT, that is now the default and the others are
   labelled explicitly as _NOIBT

 * The WINDOWS_ modifier is used as a suffix not a prefix in correctness
   statements, and likewise for windows_ in the machine code definitions.

 * The full machine code with ENDBR64 has a name ending _mc and the
   trimmed form without ENDBR64 has a name ending _tmc (t for trimmed).

* Use platform CET header where possible

The explicit ENDBR64 definition is now only used if the platform is
not already enabling CET. This improvement was suggested by lgv5 in
the discussion at https://github.com/awslabs/s2n-bignum/pull/173

* Expand ENDBR64 explanatory comment, fix whitespace

Where "fix whitespace" means both of the following:

  (1) remove all trailing spaces
  (2) expand all tabs to spaces, except in Makefiles and tools.

* Install libpcre2-dev in the CI codebuild

---------

Co-authored-by: Lucas Gabriel Vuotto <lucas@sexy.is>