fix incorrect sanity check that was added, remove too inclusive try-with
The in-range sanity check that was introduced to `get_int_le` was incorrectly failing in some cases.
This was caught by a too inclusive try-with expression in `load_elf`, causing it to
return an empty list of constant symbols when any expression in it failed.
Removed the try-with expression, and instead adopted the 'catch' function and specifically used it.