Add single-register case of TBL instruction to ARM model (#219)
* Add single-register case of TBL instruction to ARM model

This instruction uses the bytes in one argument as indices into the
bytes in the other one. Any indexing bytes >= 16 result in a zero
result for that lookup, by definition, and this is implicit in the
definition of the semantics using "word_subword". Note that our
cosimulation testing in "simulator.ml" with random numbers will have
a strong bias towards this trivial case for each lookup. However,
there's around a 50% chance that at least one of the 8 or 16 lookups
per instruction (for data sizes 64 or 128) will be nontrivial, so it
seems that it still gives adequate coverage.

* Use singleton lists in single-register TBL function

This uses a singleton list for the second argument of arm_TBL, e.g.
"arm_TBL Q29 [Q14] Q17 128", which makes the abstract syntax look
more like the ARM assembler language while also allowing for a
natural extension in the future to multi-register cases.