Support lockstep simulation of XMM_SSE registers, add READ_YMM_SSE_EQUIV to extra_word_CONV
This patch extends X86_LOCKSTEP_TAC to support SSE registers.
Previously their lockstep simulation could not be achieved for XMM_SSE
when there was no precondition on the equivalence of upper halfs of the
YMM_SSE registers, because vec operations on XMM_SSE preserves the upper halfs.
This patch relaxes X86_LOCKSTEP_TAC so that if the two output registers of
the lockstep-ping instructions are having `word_join x y` and
`word_join x' y'` as their values, and `y = y'` could be proven true,
then successfully abbreviate `y` and `y'` to the same temporary
variables.

Also, this commit has the following updates:

- Add `READ_YMM_SSE_EQUIV` to `extra_word_CONV`. This theorem rewrites
  `read YMM{n}_SSE ..` to `read YMM{n} ..`. This is a useful
  canonicalization and in fact was already existing in `X86_STEP_TAC`.\
  This patch removes the explicit `REWRITE_TAC READ_YMM_SSE_EQUIV` in
  `X86_STEP_TAC` because the `REWRITE_TAC` could rewrite post conditions
  as well.
- Improves the error message printing in `X86_CONV` and `ARM_CONV` when
  there is no `{aligned_}bytes_loaded` or `read {PC|RIP}` in
  assumptions.