Add an extra arg for data in text to mk_equiv_statement
This patch is mainly written by Yan Peng.

This adds an extra parameter to `mk_equiv_statement` that optionally
describes the constant data. The constant data is placed just after the
machine code in the text section. The output of
`define_coda_literal_from_elf` can be forwarded to these extra params.