Refresh assembly head comments (#227)
This PR follows #167 and transfers the improvements made there into the assembly head comments.

The process for doing that was, briefly: locally expand S2N_BIGNUM_STATIC in the header, reformat the entire header using clang-format, and then write a few lines of python to splat the old head comments with their replacement.

--

Some files are not fully clang-formatted, which are due to a minor modification by aqjune-aws

This commit is the work of @ctz.