Eliminate redundant or duplicated list material from HOL Light core
This avoids the use of the local function LIST_OF_FUN in favour of
the equivalent (modulo argument order) from the core, list_of_seq.
Moreover the conversions LIST_OF_SEQ_CONV, LENGTH_CONV and EL_CONV,
which are set up in the HOL Light core as of this commit

  https://github.com/jrh13/hol-light/commit/c0953a80cde3f07677c2f5e5404be08862aec9b5

are used in place of s2n-bignum versions.