Factor wordlist_from_memory out of Keccak proofs and generalize
This moves the construct "wordlist_from_memory" that is used in the
Keccak specifications back into the general background theory,
currently in common/bignum.ml. It now has a generalized word type
instead of being restricted to 64 bits, and hence can be used in
other contexts, though the nicest versions of some theorems hold
when the word size is a multiple of 8 or more generally the total
number of bits in the region is a multiple of 8. In support of
the general background, "wordlist_of_num" is also added, a natural
inverse to "num_of_wordlist".