Rename Keccak functions and move them to a separate directory
The various Keccak implementations are all currently motivated by the
use of SHA-3 in ML-KEM. But since SHA-3 and Keccak are independently
useful, it seems sensible to separate them from the other ML-KEM code.
They now live in the "sha3" subdirectory, and the function names are
prefixed by "sha3_" rather than "mlkem_".