Sharpen localization in foward and inverse NTT specs
This makes some minor stylistic changes to the way in which the
forward and inverse NTT are specified, in order to better localize
the hypothesis that the constant tables are correct. This is not
now needed as a precondition of the Hoare triple, but only as an
antecedent in the mathematical part of the postcondition. Hence
it's no longer needed to show that the function terminates and
returns.