Add collect-signatures.py that collects input/output params from C signatures
This patch adds tools/collect-signatures.py that
(1) collects input/output params from C function signatures and its comments in s2n-bignum.h
(2) stores the information at {arm,x86}/proofs/subroutine_signatures.ml .

The python script also sanity-checks whether there is any discrepancy between the comments in assembly
and s2n-bignum.h.
As a result, this commit also contains edits to a few assembly comments as well as edits to
comments in s2n-bignum.

This is the first step for upstreaming the effort of semi-automatically proving
constant-time and memory safety properties in s2n-bignum.

Also, Makefile is mildly edited to add (a subset of) dependencies from the proofs to the foundational .ml files
as well as the new subroutine_signatures.ml .