Improve SUBROUTINE_SIM_TAC tactics for subroutines with rodata
The existing [ARM|X86]_SUBROUTINE_SIM_TAC did not work satisfactorily
when the subroutine contained rodata (as is already allowed for ARM
and will be for x86 soon). Among other problems, it did not handle
the minor syntactic differences between ways of expressing relative
offsets to the pc (e.g. (pc + 4) + 8 versus pc + 12), and the style
of the tactic was not appropriate when the global routine has strictly
more rodata parameters than the subroutine. These changes make the
situation much more robust.