Use rodata section for all ARM code with constant data (#267)
* Use rodata section for curve25519_x25519base

This switches the code and proof to make use of a .rodata section for
the constant tables.

* Adopt rodata section in all curve25519 and edwards25519 code

This just updates the code, not the proofs, so far.

* Fix up code copies inside proofs for new rodata-based code

The first step in fixing up the proofs.

* Fix up preambles for new rodata proofs

This does an initial fixup of the auxiliary lemmas before the
main proofs.

* Fix up the first specs of rodata theorems and make a few other fixes

* Make first attempt to fix up remaining proofs for rodata

* More fixup of rodata proofs including final subroutine specs

* More rodata fixup for Edwards scalar mul base proofs

* Tweak table proofs in Edwards base multiplications

* More fixes to Edwards constant table proofs

Some mistaken cut-and-paste from curve25519 versions, but the
two base cases are different and the offset is different.

* More misc Edwards rodata fixups

* More small Edwards rodata fixes

* Final (?) tweaks to Edwards proofs for rodata

* Fix syntax fumbles

* Add ARM adrp instruction to allowed instruction list

* Add adrp pattern to skipped instructions for simulator

* Improve SUBROUTINE_SIM_TAC tactics for subroutines with rodata

The existing [ARM|X86]_SUBROUTINE_SIM_TAC did not work satisfactorily
when the subroutine contained rodata (as is already allowed for ARM
and will be for x86 soon). Among other problems, it did not handle
the minor syntactic differences between ways of expressing relative
offsets to the pc (e.g. (pc + 4) + 8 versus pc + 12), and the style
of the tactic was not appropriate when the global routine has strictly
more rodata parameters than the subroutine. These changes make the
situation much more robust.

* Pacify the AWS-LC delocator for the FIPS build

By enclosing some offset expressions like 96+5*16 in parentheses.

---------

Co-authored-by: Juneyoung Lee <136006969+aqjune-aws@users.noreply.github.com>