Add proof for compiler-generated fallback Keccak code
The proof is an easy adaptation of the proofs for the other variants.
Since the compiler mixes the register save/restore with the other
computation, it turned out easier just to prove a single correctness
theorem SHA3_KECCAK_F1600_ALT2_SUBROUTINE_CORRECT instead of the
usual separate proof for the "core".