Fix a few last issues with association in uglified macros
We seem forced to use -offset instead of -(offset) to make the
delocator happy. This is dangerous if "offset" expands to an
additive expression (at least for the quality of the CFI though
it doesn't affect the instructions in the code). This makes
sure we parenthesize such dangerous expressions in the macro
invocation itself.