Add call frame information (CFI) annotations (#214)
This includes call frame information annotations in all the code,
roughly following the explanation here:

  https://www.imperialviolet.org/2017/01/18/cfi.html

The change is made for both ARM and x86, largely by wrapping
certain relevant instructions in macros that augment the core
instruction with appropriate CFI directives, e.g. for ARM

#define CFI_PUSH2(lo,hi) stp     lo, hi, [sp, #-16]! ; .cfi_adjust_cfa_offset 16 ; .cfi_rel_offset lo, 0 ; .cfi_rel_offset hi, 8 

and for x86

#define CFI_PUSH(reg) push    reg ; .cfi_adjust_cfa_offset 8 ; .cfi_rel_offset reg, 0 

As part of this, two other systematic changes are made:

  - All local labels are prefixed with L to make their local-ness explicit
  - The internal header is split into separate ones for arm, x86 and x86_att

The changes were made in many stages:

* Wrap most CFI-relevant instructions up as macros

This is the start of a transformation that will annotate the code
properly with call frame information (CFI) directives, a lack in
the current s2n-bignum codebase noted in

  https://github.com/awslabs/s2n-bignum/issues/197

So far, this picks the principal CFI-relevant instructions such
as calls and returns and those that modify the stack pointer,
and simply wraps the existing instructions inside macros that
can then later be augmented with CFI directives. The definitions
of the macros in "include/_internal_s2n_bignum.h" show those
instructions that are wrapped, for ARM and/or x86:

    both    CFI_START            <marker for entry point>
    both    CFI_RET              ret

    ARM     CFI_BL(target)       bl  target
    ARM     CFI_PUSH2(lo,hi)     stp  lo, hi, [sp, #-16]!
    ARM     CFI_POP2(lo,hi)      ldp  lo, hi, [sp], #16
    ARM     CFI_INC_SP(offset)   add  sp, sp, #(offset)
    ARM     CFI_DEC_SP(offset)   sub  sp, sp, #(offset)

    x86     CFI_CALL(target)     call target
    x86     CFI_PUSH(reg)        push reg
    x86     CFI_POP(reg)         pop  reg
    x86     CFI_INC_RSP(offset)  add  rsp, offset
    x86     CFI_DEC_RSP(offset)  sub  rsp, offset

* Undo CFI macro transformations inside comments

The original script was a bit too stupid and changed these too.

* Add CFI start directives for standard ABI forms in Windows subcalls

For a relatively small number of functions, the Windows ABI form
literally calls the standard ABI form as a subroutine. For these
we add CFI_START to the start of the standard subroutine as well,
within the scope of the Windows ABI "#if ... #endif" block.

* Add CFI start directives at internal/nested procedure entry points

* Add CFI directives for non-push/pop save and restore

This adds directives for where the callee-saved registers are
directly stored onto the stack at an offset without modifying
the stack pointer. Not very many wrappers are written in this
style, but a few are.

* Add actual cfi_startproc and cfi_endproc directives

Also fix up functions with multiple exits so we only wrap the
final ret as a CFI_RET.

* Incorporate actual relative CFA offsets into CFI macros

This mostly seems to work as expected, but the combination of saving
and restoring registers in the Windows ABI form and having multiple
exit points seemed to make x86/generic/bignum_add.S need a manual
reset of the CFA offset after the first return.

* Add saved register offsets to the CFI macro definitions

* Add explicit private prefix L to all non-global labels

This fixes an issue assembling the CFI-ful files with recent Mac OS
assemblers, which objected to labels inside the CFI blocks unless
explicitly stated to be local. The error messages are not very clear,
but this seems to be the issue that is here attributed to an LLVM17
change:

  https://gitlab.winehq.org/wine/wine/-/merge_requests/4547

* Regenerate AT&T forms of CFI-enabled code

This involves trifurcating the "_internal_s2n_bignum.h" header
into three different versions with _arm, _x86 and _x86_att suffixes.
This is useful in the _x86 and _x86_att cases to distinguish the
definitions of the CFI adjustment macros, and for _arm to eliminate
irrelevant CET tests.

* Add local prefix to a rogue label that was missing it

This got missed earlier because of the space before it and then
not caught because it was only relevant for the Windows ABI.

* Update local header file name in the ARM unopt code as well

These files were just forgotten earlier. Although this code is
not part of the final build, it is used as a stepping-stone in
the proofs.

* Add a function type directive for each toplevel function

This is done via a macro S2N_BN_FUNCTION_TYPE_DIRECTIVE so that
it can be disabled on Apple, since the native Mach-O format does
not seem to use any analogous directive. Otherwise this unpacks
to the usual directive of the form (example for "bignum_mul")

  .type bignum_mul, %function

* Add a function type directive for internally called functions

This seems to matter for some versions of objdump, which can affect
the ctCheck test.

* Adopt refinement to _CET_ENDBR definition in bifurcated include files

This adopts the change https://github.com/awslabs/s2n-bignum/pull/230
but for the newly bifurcated structure of the include file(s).

* Adopt ARM-specific header with CFI macros for mlkem and sha3 code

* Add CFI directives to mlkem and sha3 ARM functions

This brings the ML-KEM related functions into line with the others
already decorated with CFI directives, as well as making a few other
stylistic tweaks to these source files for consistency (properly
prefixed labels, non-executable stack directives etc.)

* Add similar CFI annotations to new x86 Keccak function

* Avoid semicolon in new CFI macros

This copies Justin Smith's solution, aleady applied to existing macros
in 1aaa31382fd70669782594ffb59db2da84762dbc, to ensure standalone
compilability on Apple platforms without special preprocessing.

* Add function size assembler directives

This avoids some error messages from "objtool -s" and seems a
good thing to do anyway. It does not seem applicable to Mach-O.

* Localize size directives in functions with local subroutines

* Fix up size annotations related to Windows ABI subcalls

* Add size and function type annotations to local subroutines

* Add back single internal header

A temporary measure since it is still used in mlkem-native's
integration with AWS-LC.

* Simplify arguments to .cfi_adjust_cfa_offset for delocator

It seems that composite expressions of various kinds are not
accepted by the BoringSSL / AWS-LC delocator when occurring as
the argument of .cfi_adjust_cfa_offset. This simplifies arguments
in various ways to accommodate it, even at the extent of a bit
less readability and more brittleness.

* More delocator-pacifying uglification for the AWS-LC FIPS build

* Fix arithmetic fumbles

* Deglobalize constant tables, update signatures

The removal of global visibility for the rodata tables is just adopting
the recent change from the s2n-bignum mainline.

* Reduce num cores of tutorial CI to 2

* Fix a few last issues with association in uglified macros

We seem forced to use -offset instead of -(offset) to make the
delocator happy. This is dangerous if "offset" expands to an
additive expression (at least for the quality of the CFI though
it doesn't affect the instructions in the code). This makes
sure we parenthesize such dangerous expressions in the macro
invocation itself.

---------

Co-authored-by: Juneyoung Lee <136006969+aqjune-aws@users.noreply.github.com>